+++ src/index.ts
@@ -1,5 +1,7 @@
 // Export only the hook
 export * from './hooks/useIIIntegration';
 export * from './contexts/IIIntegrationContext';
-
-
+export * from './storage/AppKeyStorage';
+export * from './storage/DelegationStorage';
+export * from './storage/Ed25519KeyIdentityValueStorageWrapper';
+export * from './storage/DelegationChainValueStorageWrapper';